split main into decorators/enums
